+++ src/components/CreatorRegisterForm.jsx
@@ -25,9 +25,7 @@ const ErrorCreatorRegisterData = {
 export default function CreatorRegisterForm() {
   const [selectedImage, setSelectedImage] = useState(null);
   const [creatorData, setCreatorData] = useState(CreatorRegisterData);
-  const [errorCreatorData, setErrorCreatorData] = useState(
-    ErrorCreatorRegisterData
-  );
+  const [errorCreatorData, setErrorCreatorData] = useState(ErrorCreatorRegisterData);
 
   const handleImageChange = (event) => {
     const file = event.target.files[0];
@@ -132,10 +130,7 @@ export default function CreatorRegisterForm() {
                 Picture upload successful
               </h1>
             ) : (
-              <label
-                htmlFor="file-upload"
-                className="cursor-pointer text-center"
-              >
+              <label htmlFor="file-upload" className="cursor-pointer text-center">
                 <span
                   className={`block border-[1.5px] border-gray rounded-lg p-8 ${
                     errorCreatorData?.password ? "mt-11" : null
@@ -153,7 +148,7 @@ export default function CreatorRegisterForm() {
                 />
               </label>
             )}
-            <Button color="orange">Request Approve</Button>
+            <Button width={"full"}>Request Approve</Button>
           </div>
         </form>
       </Modal>
