+++ src/components/Input.jsx
@@ -1,5 +1,3 @@
-import React from "react";
-
 export default function Input({
   type = "text",
   placeholder,
