+++ src/features/authentication/components/LoginImage.jsx
@@ -1,0 +1,10 @@
+export default function LoginImage() {
+  return (
+    <div>
+      <img
+        src="https://idsb.tmgrup.com.tr/ly/uploads/images/2022/05/27/208313.jpg"
+        alt=""
+      />
+    </div>
+  );
+}
