+++ src/pages/Home.jsx
@@ -1,7 +1,4 @@
 import ImgCard from "../components/ImageCard";
-import ProductCommentCard from "../features/product-comment/components/ProductCommentCard";
-import ProductCommentContainer from "../features/product-comment/components/ProductCommentContainer";
-import TierContainer from "../features/tier/components/TierContainer";
 
 const mockImage =
   "https://static.thairath.co.th/media/B6FtNKtgSqRqbnNsbKEfQbPGELW2YjCcDQUpDilBzR4jVwhRbzfUfbngdYegm1bTfTXjc.webp";
@@ -14,7 +11,7 @@ const daysLeft = 23;
 export default function Home() {
   return (
     <div>
-      {/* <div className="flex justify-center mt-20">
+      <div className="flex justify-center mt-20">
         <div>
           <ImgCard
             widthSize="large"
@@ -63,8 +60,7 @@ export default function Home() {
             avatarImage={mockAvatar}
           />
         </div>
-      </div> */}
-      <TierContainer />
+      </div>
     </div>
   );
 }
